requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
boto==2.49.0
brotlipy==0.7.0
certifi==2022.5.18.1
cffi @ file:///C:/ci_310/cffi_1642682485096/work
charset-normalizer==2.0.12
colorama==0.4.4
cryptography @ file:///C:/ci/cryptography_1652101772384/work
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
Django==4.0.5
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
findspark==2.0.1
fonttools==4.33.3
future==0.18.2
h2o==3.36.1.2
h2o-pysparkling-2.1==3.32.1.7.post1
h2o-pysparkling-2.4==3.36.1.2.post1
h2o-pysparkling-3.0==3.36.1.2.post1
h2o-pysparkling-3.2==3.36.1.2.post1
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
install==1.3.5
install-jdk==0.3.0
ipykernel==6.13.1
ipython==8.4.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.3.2
jupyter-core==4.10.0
kaggle==1.5.12
kiwisolver==1.4.2
matplotlib==3.5.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.4
packaging==21.3
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
psutil==5.9.1
pure-eval==0.2.2
py4j==0.10.9.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.12.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///C:/ci/pysocks_1605287845585/work
pyspark==3.2.1
pysparkling==0.6.1
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
pywin32==304
pyzmq==23.1.0
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
spark==0.2.1
sqlparse==0.4.2
stack-data==0.2.0
tabulate==0.8.9
text-unidecode==1.3
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
tzdata==2022.1
urllib3 @ file:///C:/ci/urllib3_1650640043075/work
wcwidth==0.2.5
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
